+++ mitra_activitypub/src/queues.rs
@@ -66,7 +66,8 @@ const JOB_TIMEOUT: u32 = 3600; // 1 hour
 #[derive(Deserialize, Serialize)]
 pub struct IncomingActivityJobData {
     activity: JsonValue,
-    recipient_id: Option<String>,
+    recipient_id: Option<String>, // only when delivered to inbox
+    sender_id: Option<String>, // only when delivered to inbox
     is_authenticated: bool,
     failure_count: u32,
 }
@@ -74,13 +75,15 @@ pub struct IncomingActivityJobData {
 impl IncomingActivityJobData {
     pub fn new(
         activity: &JsonValue,
-        maybe_recipient_id: Option<&str>,
+        maybe_delivery: Option<(&str, &str)>,
         is_authenticated: bool,
     ) -> Self {
         Self {
             activity: activity.clone(),
-            recipient_id: maybe_recipient_id
-                .map(|recipient_id| recipient_id.to_owned()),
+            recipient_id: maybe_delivery
+                .map(|(recipient_id, _)| recipient_id.to_owned()),
+            sender_id: maybe_delivery
+                .map(|(_, sender_id)| sender_id.to_owned()),
             is_authenticated,
             failure_count: 0,
         }
@@ -132,8 +135,8 @@ pub async fn process_queued_incoming_activities(
             db_client,
             &job_data.activity,
             job_data.is_authenticated,
-            false, // activity was pushed
             job_data.recipient_id.as_deref(),
+            job_data.sender_id.as_deref(),
         );
         let handler_result = match tokio::time::timeout(
             duration_max,